refactor: 캐시 이름 충돌을 정규화
