+++ src/core/resolver/apk-resolver.ts
@@ -5,7 +5,7 @@
 
 import type { OSPackageInfo, PackageDependency, Repository, OSPackageSearchResult } from '../downloaders/os-shared/types';
 import { BaseOSDependencyResolver, type DependencyResolverOptions } from '../downloaders/os-shared/base-resolver';
-import { OSCacheManager } from '../downloaders/os-shared/cache-manager';
+import { OsPackageCache } from '../downloaders/os-shared/cache-manager';
 import { ApkMetadataParser } from '../downloaders/apk';
 import { isArchitectureCompatible } from '../downloaders/os-shared/repositories';
 import { searchPackagesCommon, createResolverFactory } from './os-resolver-utils';
@@ -41,7 +41,7 @@ export class ApkDependencyResolver extends BaseOSDependencyResolver {
           this.options.abortSignal
         );
         this.parsers.set(repo.id, parser);
-        const cacheKey = OSCacheManager.createKey(
+        const cacheKey = OsPackageCache.createKey(
           'apk',
           repo,
           this.options.architecture,
